fix(mullvadvpn-np): change checkver to use the Mullvad API instead of GitHub
The GitHub repo for the Mullvad VPN app is for all platforms and the
different platforms can have separate releases, updating what "latest"
points to (e.g. `android/2024.4`), leading checkver to fail. This commit
changes checkver to use the Mullvad API to query directly for the latest
version for Windows.